engine-py/requirements.txt
# Smart Workflow Automation Tool - Python AI Engine
# Step 2: AI Core with Gemini

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data validation
pydantic==2.5.3

# HTTP client
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.8.3
